Removing the version number from the GRAYLOG2CTL_DIR variable in init.d-graylog2-web-interface Removing the DAEMON_OPTS variable also.
If not overwritten the old value for GRAYLOG2CTL_DIR (/opt/graylog2-web-interface-preview.1) will be used
and since that directory doesn't exist in this RPM installation the server will refuse to start.

The DAEMON_OPTS variable is unused, testing proved to be problematic and the values for this variable are
required for the application to run. They are now hard coded into the startup command.