Managing a new system user and group graylog2 in graylog2-server.spec
At the pre stage the system group graylog2 is created if it doesn't already exist, same with the
user.
At the preun stage this user and group are removed if they aren't already, using getent.

This new user/group will maintain ownership of the log directory, as write permissions are required,
and the install root /opt/graylog2-server/