Maintaining the new system user and group graylog2 in graylog2-web-interface.spec
In the pre stage the graylog2 group is created, if it hasn't already (using getent), then the
user is created, if it hasn't already.

The graylog2 user/group will be the new owner of the log directory and everything under /opt/graylog2-web-interface

CONCERN: If this package is installed on the same server as the graylog2-server package and one of these
packages is removed the other may not work properly. May need to consider combining the spec files.
At the preun stage this new user and group is removed.