Adding the lib_dir and CLASSPATH variables to sysconfig-graylog2-web-interface Running the graylog2-web-interface as the root user and breaking the startup command into multiple lines in init.d-graylog2-web-interface
The additional variables in sysconfig-graylog2-web-interface lightens the complete dependency on
variables declared in init.d-graylog2-web-interface. These variables are not likely to change.

In an attempt to run the web-interface as an unpriviledged user, configured in init.d-graylog2-web-interface,
resulted in the server not being able to create the syslog inputs. Running it as the root user is now the
permanent workaround.

Breaking the startup command arguments into multiple, escaped lines adds readability and allows for bulk
insert and removal of different arguments.