Running the graylog2-server as root user from unpriviledged graylog2 user in init.d-graylog2-server
While starting up the server as the graylog2 user creating the log inputs failed to bind the ports
because the user didn't have permissions to do so. May be able to work around this in the future,
but for now running it as the root user ensures all functionality is present.