Adding the log4j configuration file available on https://github.com/Graylog2/graylog2-server
https://github.com/corey-hammerton/graylog2-rpms/issues/2

Will first add the unmodified log4j.xml file from the graylog2-server project to the SOURCE
directory. This file will be unused until such a time when documentation for externalizing
the log4j configurations for the graylog2-server and web-interface components.

In the meantime minor modifications to this file may be committed before it's included
in the init scripts and spec files to be built in the rpms.