Adding new file log4j-graylog2-server.xml
https://github.com/corey-hammerton/graylog2-rpms/issues/2

The new file is an externalized version of the log4j.xml file that
comes pre-packaged in the graylog2-server package. In my experience
as a sysadmin this is only good for development/test environments,
come to production environments the sysadmins need control of the
logging. To monitor the application status and send log output to
the developers to investigate their code in order to fix any bugs or
other issues.

Will configure my test graylog2-server instance with a daily
rolling file appender to write to a self-managed log file instead
of to the currently configured stdout and redirecting to a file.