Removing the sleep command from the start function of SOURCES/init.d-graylog2-server
https://github.com/corey-hammerton/graylog2-rpms/issues/1

Since the native starup process of the graylog2 package family has
improved during the preview and now release candidate releases there
is no reason to have port testing or waiting at the init script
startup level. So removing the sleep command from the start function.

Adding a sleep command to the restart function. When attempting to
run a simple stop and start within the script the logs state that
a required port is already in use. By adding a sleep command, this
waits for 5 seconds which may be a little excessive, to the restart
will give the java service a chance to shutdown cleanly before
attempting to start it up again.

At the beginning of the script, if the java command is not found
exit the script with an exit code of 1. With a recent java upgrade
and I forgot to update the configuration file with the new path
before restarting the system the init script did see that the JAVA_CMD
path was not executable and displayed a failure, but still continued.
The exit with code 1 will prevent future issues of this sort.