Add new file log4j-graylog2-web-interface.xml
https://github.com/corey-hammerton/graylog2-rpms/issues/2

As what was completed for the graylog2-server package, externalizing
the log4j configuration for the graylog2-web-interface must also
be completed. The contents of this file were taken from the
graylog2-web-interface documentation and modified to that the
behaviour mimics the graylog2-server configuration, even though the
classes here are different.

Initial testing for this configuration file needs to be completed
before I will include this file in the SPEC file.