Adding the log4j configuration for the web-interface to the main tree
Adding a sleep command in the result function of
SOURCES/init.d-graylog2-web-interface, because running a simple
stop and restart causing an exception because a port is already
in use because the services hasn't shut down yet before a start
was attempted.

Exiting with code 1 if the JAVA_CMD variable is not executable in
SOURCES/init.d-graylog2-web-interface

Removing the file redirection from the start command in
SOURCES/init.d-graylog2-web-interface. Since the new log4j config
no longer outputs to System.out/stdout a file redirection is not
needed.

Removing the commented classpath variable from
SOURCES/sysconfig-graylog2-web-interface, it contains the list of
individual jar files in the lib directory. Making this file another
thing to update on version updates, besides, any problems that may
occur for users will be based on their setup.

Including and installing the log4j-graylog2-web-interface.xml file
from SPECS/graylog2-web-interface.spec under /etc/graylog2/.
Officially adding it to the rpm building process.