Updating packages to 0.20.0-rc.1-1
Changing the release number of graylog2-server.spec and
graylog2-web-interface.spec to rc.1-1 and adding a changelog entry
for the new release.

Not 100% sure if an RPM build will work since the relase number
now contains a dash. May result in a custom build specific only
to this release.