Reverting the release number back to rc.1
Since the rc.1-1 release is a special case because of the dash
in the release number a standard increment in the release number
in SPECS/graylog2-server.spec and SPECS/graylog2-web-interface.spec
cannot be completed. The RPM builds fail because of illegal char
'-'.

In both files, resetting the Release number, removing the changelog
entry and updating the Source0 and setup entries.