Replacing all string matching the package name with the name setting.
In graylog2-server.spec all matching "graylog2-server" strings, in
the Sources list and install step are replaced with %{name}.

Same in graylog2-web-interface.spec with "graylog2-web-interface"

The spec files are as a result smaller and more portable. Allowing
for another related package, for instance graylog2-radio, to be
created as a copy if an existing spec. Fewer changes will have
to be made as a result.