Adding rundamentary configtest function
https://github.com/corey-hammerton/graylog2-rpms/issues/5

Files Modified:
SOURCES/init.d-graylog2-server

Adding a configtest function to SOURCES/init.d-graylog2-server
to test the syntax of the configuration file.

Right now this is just a simple function. I would like to validate
the log4j configuration but in order to do that will have to
implement said configuration differently.