Adding a configtest function in the web-interface init-script.
https://github.com/corey-hammerton/graylog2-rpms/issues/5

Files Modified:
SOURCES/init.d-graylog2-web-interface
SOURCES/sysconfig-graylog2-web-interface
SPECS/graylog2-web-interface.spec

Adding a new variable GRAYLOG2_LOG_CONF in
SOURCES/init.d-graylog2-web-interface and
SOURCES/sysconfig-graylog2-web-interface which eliminates the need
to hard-code the settings in the ADDITIONAL_JAVA_OPTS variable.
This is the only possible way to test the syntax of the log4j.xml
configuration.

Adding the configtest function to SOURCES/init.d-graylog2-web-interface
which only runs xmllint on the GRAYLOG2_LOG_CONF file if it
exists. Since the web-interface doesn't have configtest
functionality we cannot check the functionality of the web-interface
configuration.

Adding a dependency for libxml2 in SPECS/graylog2-web-interface.spec.
In order to be able to run xmllint.